Move executor pod construction to a separate class.
This is the first of several measures to make
KubernetesClusterSchedulerBackend feasible to test.
